backend/requirements.txt
fastapi
uvicorn
websockets
deepgram-sdk
python-multipart
